refactor(generator): pull contract orchestration into ContractFinder.Discover*
Move the Step 2 (contracts-assembly map), Step 3 (interface scan), and Step 3b
(implementation scan) blocks from SymbolDiscovery.Extract into two new orchestration
methods: ContractFinder.BuildContractsAssemblyMap and
ContractFinder.DiscoverInterfacesAndImplementations. The convention-DTO scan block
between them remains in Extract.
